requirement.txt

torch
transformer

